Fix control flow for interval/wait action; remove actionList and replace with single currentAction
Agent.actionList was used for queueing manually triggered actions, but now that the actions are chained with conditions, only the first item in the list is ordinarily used and it is making control flow unnecessarily complicated, so can be replaced with a single action reference.

The ability to manually "pilot" a selected agent with a queue of actions triggered with buttons is still useful, but this can be implemented again later within the current framework, e.g. a user creates a list of action 1 and action 2 in the UI, and the business logic creates action 1 with a transition to action 2 using the preset 'isComplete' condition.