more schema changes and fixing incomplete code for Schelling model
- interval actions now have a 'intervalType' setting, which can be either
1. frames: a fixed frame duration (currently used) or
2. untilNextInterval: wait until a given frame e.g. next day (not yet implemented)

Meaning that if agents finish the same task at different times, they can sync up
again for next action

- create and delete actions for agents during the runtime of the simulation (ie.
when 'scene' is playing) no longer save to the database. This is so that a scene
can be set up and then reset to run again by reloading the page. Agent CRUD
when scene not playing are saved to database.